Home

Slacker.jl

Slacker.addConfig

Slacker.loadConfig

Slacker.readSettingsFile

Slacker.removeConfigFile

Slacker.sendSlackMessage

Slacker.sendSlackMessage

Slacker.addConfig — Function.

addConfig(config::SlackConfig, name::String = "default")

Add a Slack server configuration to the registry. Multiple servers can be defined by adjusting the name parameter.

source

Slacker.loadConfig — Function.

loadConfig(name="default")

Load the default configuration from the registry. Other configurations may be loaded by specifying the name parameter.

source

Slacker.readSettingsFile — Method.

readSettingsFile()

Load the registry and return all registered servers as a Dict.

source

Slacker.removeConfigFile — Method.

removeConfigFile()

Deletes the whole registry from the computer.

source

Slacker.sendSlackMessage — Function.

sendSlackMessage(text, cfg::String = "default")

Sends a message textto the configured default slack server or the one specified in cfg.

source

Slacker.sendSlackMessage — Method.

sendSlackMessage(text, cfg::SlackConfig)

Sends a message textto Server defined in the configuration cfg.

source