Home

Slacker.jl

Index

Slacker.addConfig

Slacker.hasConfig

Slacker.loadConfig

Slacker.readConfigFile

Slacker.removeConfigFile

Slacker.sendMessage

Slacker.sendMessage

Slacker.jl is a package that allows sending messages to Slack. It was designed to in form a group of users, when a computation has finished.

Functions

Slacker.addConfig — Function.

Add a Slack server configuration to the registry. Multiple servers can be defined by adjusting the name parameter.

Examples

julia> addConfig(SlackConfig("server.url", "username", "#channel", ":smile:"))
1

source

Slacker.hasConfig — Method.

Tests whether a configuration exists

source

Slacker.loadConfig — Function.

Load the default configuration from the registry. Other configurations may be loaded by specifying the name parameter.

source

Slacker.readConfigFile — Method.

Load the registry and return all registered servers as a Dict.

source

Slacker.removeConfigFile — Method.

Deletes the whole registry from the computer.

source

Slacker.sendMessage — Function.

Sends a message textto the configured default slack server or the one specified in cfg.

source

Slacker.sendMessage — Method.

Sends a message textto Server defined in the configuration cfg.

source